LayoutTests/fast/regions/auto-height/last-region-auto-height-no-breaks.html

<!doctype html>
<html>
    <head>
        <!--
        Test for regions auto height prototype. The last region has auto height.
        -->
        <script src="../../../fast/js/resources/js-test-pre.js"></script>
        <style>
            body { font-family: monospace; }
            #article { -webkit-flow-into: flow; }
            #region { -webkit-flow-from: flow; width: 200px; height: 50px; border: 1px solid red; }
            #region2 { -webkit-flow-from: flow; width: 100px; border: 1px solid red; }
        </style>
    </head>
    <body>
        <div id="article">
            <p id="p1">Text inside article p1. Text inside article p1.</p>
            <p id="p2">Text inside article p2. Text inside article p2.</p>
        </div>
        <div id="region"></div>
        <div id="region2"></div>
        <script>
            if (window.layoutTestController)
                window.layoutTestController.dumpAsText();
            description("Test for region auto height prototype.");
            var namedFlow = document.webkitGetFlowByName("flow");
            var p1RegionList = namedFlow.getRegionsByContentNode(document.getElementById("p1"));
            var p2RegionList = namedFlow.getRegionsByContentNode(document.getElementById("p2"));

            shouldEvaluateTo("p1RegionList.length", 1);
            shouldBeEqualToString("p1RegionList.item(0).id", "region");
            shouldEvaluateTo("p2RegionList.length", 1);
            shouldBeEqualToString("p2RegionList.item(0).id", "region2");

            document.getElementById("article").style.visibility = "hidden";
            document.getElementById("region").style.visibility = "hidden";
            document.getElementById("region2").style.visibility = "hidden";
        </script>
    </body>
    <script src="../../../fast/js/resources/js-test-post.js"></script>
</html>
